requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.7.0
pydantic-settings==2.4.0
psycopg2-binary==2.9.7
python-multipart==0.0.6
sse-starlette==1.6.1
joblib==1.4.2
google-generativeai==0.8.5
httpx>=0.25
openai>=1.14.0
sentence-transformers
tqdm>=4.66.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.14.0
tenacity
